added first project files
