+++ src/engine.h
@@ -1,0 +1,10 @@
+// snake engine
+// licensed under MIT
+// by MCWertGaming
+
+namespace engine
+{
+    unsigned int consoleHight();
+    unsigned int consoleWidth();
+    void clearConsole();
+}
