+++ src/snake.cpp
@@ -1,0 +1,3 @@
+// snake class source
+
+#include "snake.h"
