+++ src/snake.h
@@ -1,0 +1,31 @@
+// snake game components header
+// licensed under MIT
+// source is seperated!
+
+namespace snake
+{
+    class snake
+    {
+    private:
+        // stores all game components - field[hight][width]
+        unsigned int field[50][20];
+        /*
+         * Data is represented in snake::id
+         */
+    public:
+        /* parsing functions */
+
+        /* game manipulation */
+
+        /* game control */
+
+        /* frontend parsing */
+
+    };
+    namespace id
+    {
+        unsigned int snakeBody();
+        unsigned int snakeHead();
+        unsigned int apple();
+    }
+}
