+++ src/snakeAdditions.cpp
@@ -1,0 +1,16 @@
+// additional snake namespace source
+// licensed under MIT
+#include "snake.h"
+
+unsigned int snake::id::snakeBody()
+{
+    return 1;
+}
+unsigned int snake::id::snakeHead()
+{
+    return 2;
+}
+unsigned int snake::id::apple()
+{
+    return 3;
+}
